feat: add hero slideshow, centralize CTA link config, fix navbar button styling
