+++ src/config/links.ts
@@ -1,0 +1,6 @@
+const links = {
+  primaryCTA: 'https://forms.gle/bBGgiNmKEx1xTVJRA', 
+
+};
+
+export default links;
