+++ src/pages/Home/data/heroImages.ts
@@ -1,0 +1,16 @@
+import image1 from '@/assets/images/open1.jpg'
+import image2 from '@/assets/images/open2.jpg'
+import image3 from '@/assets/images/open3.jpg'
+import image4 from '@/assets/images/open4.jpg'
+import image5 from '@/assets/images/open5.jpg'
+import image6 from '@/assets/images/open6.jpg'
+import image7 from '@/assets/images/open7.jpg'
+import image8 from '@/assets/images/open8.jpg'
+import image9 from '@/assets/images/open9.jpg'
+import image12 from '@/assets/images/open12.jpg'
+
+const heroImages = [
+    image1,image2,image3,image4, image5,image6, image7, image8, image9, image12
+]
+
+export default heroImages;
